Beginning to flesh out structures for input rules
More info on input rules, they will act as a wrapper for the function and make sure the data is valid and (SQL) safe to provide to the function to return data, with the end goal of reducing that work for the programmer. I hope to eventually bring this to make easier SQL queries.